BuddyPress: nth attempt at proper unhooking
This commit has a new set of directions to deactivate/unhook exclusive
components for the current and displayed user.

1. Determine vgsr-ship from custom member-type query
In order to determine whether the current `is_user_vgsr()` before the
member-types are registered (at bp_init), we now use a custom query that
queries all user ids attached to the member-type taxonomy term(s). This
array is used for checking the user status up to the moment when the
member-types are registered.

2. Active components for the current user
Based on user logged-in status or `is_user_vgsr()` (see 1.) the list of
active components is emptied or modified for exclusive components
respectively. Since 1. we can now modify this list before any component is
actually loaded, which removes the need to filter `bp_is_active()`.

3. Unhook displayed user nav items
Exclusive BP nav items cannot just be removed for the displayed user,
since they're also used for the current user in other places. That is why
we now manually overwrite their `show_for_displayed_user` param, and
unhook the screen functions, also manually. Consequently BP does a 404
when you'd attempt to visit a displayed user's profile page.

4. Unhook bp-legacy theme component actions
When using theme-compat, the `BP_Legacy` class inserts various buttons
and other logic based on component activeness. Since we cannot filter
these out in `bp_is_active()` (because that would also apply for the
current user), they are manually unhooked where possible.

Mind that some hooks or other custom plugin logic for the displayed user
is not covered by this, because the global state of components is defined
for the current user.