Event Organiser: introduce Events home page
This change enables displaying a custom page or template page to be used
for the main events archive page instead of a simple events archive.
Supporting this, the Upcoming Events widget is introduced which displays
all events for the upcoming n days, listed by day.

The settings option for selecting a Page to be used as the Events home
page was already added in c25b390. When no page is selected, the plugin
defaults to the first template that is provided as a fallback, either
provided by the current theme or the plugin's own template files. This is
called using EO's own `eo_locate_template()` templating logic.